@TestConfiguration moved to BaseTestContract and BaseTestIntegration as was conflicting with smoke tests
