+++ src/integrationTest/java/uk/gov/hmcts/reform/orgrolemapping/controller/BaseTestIntegration.java
@@ -1,10 +1,33 @@
 package uk.gov.hmcts.reform.orgrolemapping.controller;
 
 import com.azure.messaging.servicebus.ServiceBusSenderClient;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.opentable.db.postgres.embedded.EmbeddedPostgres;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.context.ApplicationContextInitializer;
+import org.springframework.context.ApplicationListener;
+import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.event.ContextClosedEvent;
+import org.springframework.jdbc.datasource.SingleConnectionDataSource;
+import org.springframework.lang.NonNull;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.support.TestPropertySourceUtils;
+import uk.gov.hmcts.reform.orgrolemapping.controller.utils.WiremockFixtures;
 
+import javax.annotation.PreDestroy;
+import javax.sql.DataSource;
+import java.io.IOException;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+import java.util.Properties;
+
+@ContextConfiguration(initializers = {BaseTestIntegration.WireMockServerInitializer.class})
 @ActiveProfiles("itest")
 public abstract class BaseTestIntegration extends BaseTest {
 
@@ -15,4 +38,62 @@ public abstract class BaseTestIntegration extends BaseTest {
     @MockBean
     @Qualifier("jrdPublisher")
     ServiceBusSenderClient serviceBusSenderClientJrd;
+
+    @TestConfiguration
+    static class Configuration {
+        Connection connection;
+
+        @Bean
+        public EmbeddedPostgres embeddedPostgres() throws IOException {
+            return EmbeddedPostgres
+                    .builder()
+                    .start();
+        }
+
+        @Bean
+        public DataSource dataSource(@Autowired EmbeddedPostgres pg) throws Exception {
+
+            final Properties props = new Properties();
+            // Instruct JDBC to accept JSON string for JSONB
+            props.setProperty("stringtype", "unspecified");
+            props.setProperty("user", "postgres");
+            connection = DriverManager.getConnection(pg.getJdbcUrl("postgres"), props);
+            return new SingleConnectionDataSource(connection, true);
+        }
+
+
+        @PreDestroy
+        public void contextDestroyed() throws SQLException {
+            if (connection != null) {
+                connection.close();
+            }
+        }
+    }
+
+    public static class WireMockServerInitializer
+            implements ApplicationContextInitializer<ConfigurableApplicationContext> {
+
+        private final WiremockFixtures wiremockFixtures = new WiremockFixtures();
+
+        @Override
+        public void initialize(@NonNull ConfigurableApplicationContext applicationContext) {
+
+            TestPropertySourceUtils.addInlinedPropertiesToEnvironment(
+                    applicationContext,
+                    "wiremock.server.port=" + WIRE_MOCK_SERVER.port()
+            );
+
+            try {
+                wiremockFixtures.stubIdamConfig();
+            } catch (JsonProcessingException e) {
+                throw new RuntimeException(e);
+            }
+
+            applicationContext.addApplicationListener((ApplicationListener<ContextClosedEvent>) event -> {
+                if (WIRE_MOCK_SERVER.isRunning()) {
+                    WIRE_MOCK_SERVER.shutdown();
+                }
+            });
+        }
+    }
 }
